moves to own files and adds real world examples
