+++ test-d/entries.ts
@@ -1,4 +1,5 @@
-import {Entry, Entries} from '../source/entries';
+import { Entries} from '../source/entries';
+import { Entry } from '../source/entry';
 import {expectAssignable} from 'tsd';
 
 // Objects
